fix: rename registration page folder, remove unused css property
