+++ src/pages/productPage/ProductPage.css
@@ -13,7 +13,6 @@
   display: flex;
   flex-direction: column;
   padding-right: 10px;
-  width: '50%';
 }
 
 .ant-image .product-img {
